Example Segmentation Code
This code adds evenly spaced points to a dataframe of coors from a polyline.

Options to include orignal point and to add a final point

Code does not use spatial packages,
Does not calculate geospatial distance
direction vector instead of bearing/heading used.

To be useful needs to identifty points as orignal or added.